open-models-train/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
accelerate==0.28.0
datasets==2.14.5
deepspeed==0.14.0
sentencepiece==0.2.0
torch==2.2.1+cu121
transformers==4.38.2
tqdm==4.66.1
PyYAML==6.0.1
tensorboard==2.15.1
